refactor(shared): Split errors and HTTP response modules
