+++ src/shared/errors/map-error-to-http.ts
@@ -6,7 +6,8 @@
  * This module is the single bridge between the {@link BaseError} hierarchy and HTTP API responses.
  * Route wrappers such as {@link withErrorHandling} and {@link withZodValidation} rely on
  * {@link mapErrorToHttp} (directly or via {@link createErrorResponse}) so clients always receive
- * a consistent `{ code, message, meta }` error body shape.
+ * a consistent `{ code, message, meta }` error body shape. Per-class mapping helpers live in
+ * {@link error-http-maps}; payload types in {@link http-error-types}.
  *
  * **HTTP status mapping overview**
  *
@@ -39,149 +40,19 @@
 
 import {
   AuthError,
-  BaseError,
   DomainError,
   InfraError,
   ValidationError,
 } from '@shared/errors/app-error'
-import { ErrorCodes, type ErrorCode } from '@shared/errors/codes'
+import { ErrorCodes } from '@shared/errors/codes'
+import type { HttpErrorResponse } from '@shared/errors/http-error-types'
+import {
+  buildAppErrorBody,
+  mapAuthErrorToHttp,
+  mapDomainErrorToHttp,
+} from '@shared/errors/error-http-maps'
 import { logger } from '@utils/logger-util'
 import * as SentryNode from '@sentry/node'
-
-/**
- * JSON-serializable error payload returned in HTTP responses.
- *
- * @remarks
- * Serialized as the `body` half of {@link HttpErrorResponse}. The `meta.details` field mirrors
- * `BaseError.details` when present; it may contain Zod issue arrays, entity identifiers, or
- * operation names depending on which factory or class created the error.
- */
-type HttpErrorBody = {
-  /** Stable application error code (e.g. `ANIME_NOT_FOUND`). */
-  code: string
-  /** Human-readable message safe to expose to API clients (may be generic for 500 responses). */
-  message: string
-  /** Optional wrapper; `details` holds structured context from the originating error. */
-  meta?: Record<string, unknown>
-}
-
-/**
- * HTTP status plus structured error body produced by the mapper.
- */
-type HttpErrorResponse = {
-  /** HTTP status code (400, 401, 403, 404, or 500). */
-  status: number
-  /** JSON error body with `code`, `message`, and optional `meta`. */
-  body: HttpErrorBody
-}
-
-/**
- * Auth error codes that map to **401 Unauthorized**.
- *
- * @remarks
- * Used by {@link mapAuthErrorToHttp}. All other recognized {@link AuthError} codes fall through
- * to {@link ErrorCodes.AUTH_FORBIDDEN} (403) when explicitly matched.
- *
- * @internal
- */
-const UNAUTHORIZED_AUTH_CODES = new Set<ErrorCode>([
-  ErrorCodes.AUTH_REQUIRED,
-  ErrorCodes.AUTH_INVALID_TOKEN,
-  ErrorCodes.AUTH_SESSION_EXPIRED,
-])
-
-/**
- * Domain error codes that map to **404 Not Found**.
- *
- * @remarks
- * Any other {@link DomainError} code maps to **400 Bad Request** and is logged at `error` severity.
- *
- * @internal
- */
-const NOT_FOUND_DOMAIN_CODES = new Set<ErrorCode>([
-  ErrorCodes.ANIME_NOT_FOUND,
-  ErrorCodes.MUSIC_NOT_FOUND,
-  ErrorCodes.WATCH_STATE_NOT_FOUND,
-  ErrorCodes.COLLECTION_NOT_FOUND,
-  ErrorCodes.USER_NOT_FOUND,
-])
-
-/**
- * Builds the standard HTTP error body from a {@link BaseError} instance.
- *
- * @param error - Application error with `code`, `message`, and optional `details`
- * @returns Body where `meta.details` echoes `error.details` when defined
- *
- * @internal
- */
-function buildAppErrorBody(error: BaseError): HttpErrorBody {
-  return {
-    code: error.code,
-    message: error.message,
-    meta: { details: error.details },
-  }
-}
-
-/**
- * Maps {@link AuthError} instances to 401 or 403 responses.
- *
- * @param error - Authentication or authorization failure
- * @returns HTTP response for recognized auth codes, or `undefined` if the code is not handled here
- *
- * @remarks
- * - **401**: `AUTH_REQUIRED`, `AUTH_INVALID_TOKEN`, `AUTH_SESSION_EXPIRED` — logged at `warn`.
- * - **403**: `AUTH_FORBIDDEN` — logged at `warn`.
- * - Unrecognized {@link AuthError} codes return `undefined` so {@link mapErrorToHttp} can fall through
- *   to generic unknown-error handling.
- *
- * @internal
- */
-function mapAuthErrorToHttp(error: AuthError): HttpErrorResponse | undefined {
-  if (UNAUTHORIZED_AUTH_CODES.has(error.code)) {
-    logger.warn({ err: error }, 'Auth error - unauthorized')
-    return {
-      status: 401,
-      body: buildAppErrorBody(error),
-    }
-  }
-
-  if (error.code === ErrorCodes.AUTH_FORBIDDEN) {
-    logger.warn({ err: error }, 'Auth error - forbidden')
-    return {
-      status: 403,
-      body: buildAppErrorBody(error),
-    }
-  }
-
-  return undefined
-}
-
-/**
- * Maps {@link DomainError} instances to 404 or 400 responses.
- *
- * @param error - Business-rule violation from domain logic
- * @returns **404** when `error.code` is in {@link NOT_FOUND_DOMAIN_CODES}; otherwise **400**
- *
- * @remarks
- * Not-found codes are logged at `warn`; other domain errors at `error`. Neither path reports to Sentry.
- *
- * @internal
- */
-function mapDomainErrorToHttp(error: DomainError): HttpErrorResponse {
-  if (NOT_FOUND_DOMAIN_CODES.has(error.code)) {
-    logger.warn({ err: error }, 'Domain error - not found')
-    return {
-      status: 404,
-      body: buildAppErrorBody(error),
-    }
-  }
-
-    logger.error({ err: error }, 'Domain error')
-  return {
-    status: 400,
-    body: buildAppErrorBody(error),
-  }
-}
 
 /**
  * Converts a thrown value into an HTTP status and structured JSON error body.
